BVMU-68 refactor: add authorization annotation to security to do with updating school status
